Refactor BillDetailForm component and sub-components
BillDetailForm is responsible for maintaining the state of the form. A
one way data flow ensures that the state in the BillDetailForm is the
single source of truth.

The ParticipantForm is responsible for acquiring information about a
single participant and uses the onChange handler from the BillDetailForm
to update the participant state.

Similarly, the ItemForm is responsible for acquiring information about a
single item and uses the onChange handler from it's parent, the
ParticipantForm.