Add feature: delete bill
A bill can be deleted from the 'Read Bill Screen'. Before a bill is
deleted, a confirmation message is shown to the users. On deletion, the
user is redirected to the home screen. The 'Read Bill Screen' will only
render for bills that existed. When the screen is accessed for a deleted
bill, the user is automatically redirected to the home screen